requirements.txt
python-magic==0.4.18
pexpect==4.8.0
ubi-reader==0.6.7
python-lzo==1.12
six==1.15.0